backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
faiss-cpu
python-dotenv
sentence-transformers
